requirement.txt
tensorflow-cpu
streamlit
gdown
pillow
numpy